% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{nord}
\alias{nord}
\title{nord palette colors ramped to a specified length}
\usage{
nord(palette = "polarnight", n, alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'nord_palettes' list}

\item{n}{Number of colors to display}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
nord palette colors ramped to a specified length
}
\examples{
library(nord)
image(volcano, col = nord("aurora", 20))
}
