% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{nord_pal}
\alias{nord_pal}
\title{nord palette with ramped colours}
\usage{
nord_pal(palette = "polarnight", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'nord_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
nord palette with ramped colours
}
