% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{nord_palettes}
\alias{nord_palettes}
\title{Nord color palettes for R from ArcticIceStudios https://git.io/nord}
\format{An object of class \code{list} of length 6.}
\usage{
nord_palettes
}
\description{
An arctic, north-bluish color palette.
Created for the clean- and minimal flat design pattern to achieve a optimal focus and readability for code syntax highlighting and UI. It consists of four palettes utilzing a total of sixteen, carefully selected, dimmed pastel colors for a eye-comfortable, but yet colorful ambiance.
}
\details{
The available palettes are:
polarnight
snowstorm
frost
aurora
}
\examples{

}
\keyword{datasets}
