% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{nord_show_palette}
\alias{nord_show_palette}
\title{Display a color palette}
\usage{
nord_show_palette(palette)
}
\arguments{
\item{palette}{vector of character hex RGB values}
}
\description{
Given a character vector (hex RGB values), display palette in graphics window.
}
\examples{
library(nord)
nord_show_palette("halifax_harbor")
}
