% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_nord}
\alias{scale_fill_nord}
\title{#' nord fill scale for ggplot2}
\usage{
scale_fill_nord(palette = "polarnight", discrete = TRUE, alpha = 1,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose from 'nord_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colours is reversed.}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
#' nord fill scale for ggplot2
}
\examples{
library(ggplot2)
library(nord)

ggplot(diamonds) +
  geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_nord("victory_bonds")
}
